+++ ft_printf_tmax530/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: yanis <yanis@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/22 22:29:25 by yanis             #+#    #+#             */
-/*   Updated: 2025/09/23 21:37:23 by yanis            ###   ########.fr       */
+/*   Updated: 2025/10/22 01:54:02 by yanis            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 int	ft_putnbr(int nb, int *i);
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
-int	ft_puthex(unsigned int nb, char c);
+int	ft_puthex(unsigned long nb, char c);
 int	ft_putptr(void *ptr);
 int	ft_putunbr(unsigned int nb);
 int	ft_printf(const char *str, ...);
